+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/webextension/GeckoWebExtensionException.kt
@@ -37,6 +37,7 @@ class GeckoWebExtensionException(throwable: Throwable) : WebExtensionException(t
                     ERROR_BLOCKLISTED -> WebExtensionInstallException.Blocklisted(
                         extensionId = throwable.extensionId,
                         extensionName = throwable.extensionName,
+                        extensionVersion = throwable.extensionVersion,
                         throwable,
                     )
 
@@ -70,6 +71,7 @@ class GeckoWebExtensionException(throwable: Throwable) : WebExtensionException(t
                     ERROR_SOFT_BLOCKED -> WebExtensionInstallException.SoftBlocked(
                         extensionId = throwable.extensionId,
                         extensionName = throwable.extensionName,
+                        extensionVersion = throwable.extensionVersion,
                         throwable,
                     )